Refactor project to support Python 2.7.15
