+++ .gitignore
@@ -9,3 +9,4 @@ dist
 doc/reference/build/*
 lib/*.egg-info
 *.sav
+.python-version
