+++ README.md
@@ -1,0 +1,34 @@
+
+etk.docking
+===============================
+
+version number: 0.3
+author: Dieter Verfaillie
+
+Overview
+--------
+
+PyGObject Docking Widgets
+
+Installation / Usage
+--------------------
+
+To install use pip:
+
+    $ pip install etk.docking
+
+
+Or clone the repo:
+
+    $ git clone https://github.com/dieterv/etk_docking.git
+    $ python setup.py install
+
+Contributing
+------------
+
+TBD
+
+Example
+-------
+
+TBD
