+++ docs/examples/demo.py
