+++ docs/examples/dockgroupdemo.py
