+++ docs/platform/colors.html
