+++ docs/platform/colors.py
