+++ docs/reference/source/dev/serialization.rst
