+++ etkdocking/__init__.py
@@ -30,6 +30,12 @@ __docformat__ = 'restructuredtext'
 ############################################################################
 # Initialization
 ############################################################################
+try:
+    import pygtk
+except ImportError:
+    pass
+else:
+    pygtk.require('2.0')
 import os, gtk
 
 # Register some custom icons into the default icon theme
@@ -62,10 +68,8 @@ del os, gtk, path
 # This also defines the widgets that can be considered public.
 ############################################################################
 from .docklayout import DockLayout
-from .docklayout import add_new_group_left, add_new_group_right, \
-                        add_new_group_above, add_new_group_below, add_new_group_floating
-from .dockframe import DockFrame
+from etkdocking.dockframe import DockFrame
 from .dockpaned import DockPaned
-from .dockgroup import DockGroup
+from etkdocking.dockgroup import DockGroup
 from .dockitem import DockItem
 from .docksettings import settings
