+++ etkdocking/compactbutton.py
