+++ etkdocking/icons/16x16/compact-close-prelight.png
