+++ etkdocking/icons/16x16/compact-minimize.png
