+++ etkdocking/icons/16x16/compact-restore.png
