+++ etkdocking/util.py
