+++ requirements.txt
@@ -1,0 +1,9 @@
+# Install
+simplegeneric
+
+# Dev/Deployment
+sphinx
+sphinx_rtd_theme
+nose
+coverage
+pypi-publisher
