+++ setup.cfg
@@ -1,17 +1,15 @@
-[egg_info]
-tag_build = .dev
-tag_date = 0
-tag_svn_revision = 0
+[bdist_wheel]
+universal=1
+
+[metadata]
+description-file=README.md
 
 [nosetests]
-tests=etk.docking,doc
-with-doctest=0
+tests=tests,docs
 verbosity=0
 detailed-errors=1
-with-coverage=1
-cover-package=etk.docking
 
 [build_sphinx]
-source-dir = doc/reference/source/
-build-dir = doc/reference/build/
+source-dir = docs/reference/source/
+build-dir = docs/reference/build/
 all_files = 1
